Use default values for client and context if client not provided
A resource server can use a userInfoUri endpoint to decode a token
without having a clientId, so we need to take that into account.
It was already working in fact, but the OAuth2RestOperations was
being created as needed, which meant that the user-supplied
UserInfoRestTemplateCustomizer callbacks were not being applied to
it.

Fixes gh-50